chore(error-tracking): inline rule modal logics, remove factory
Replace the factory-built rule modal logics with standard kea() calls
so kea-typegen can auto-generate the LogicType files instead of
requiring manually maintained type definitions.
